authentication/requirement.txt
asgiref==3.3.4
astroid==2.5.6
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
Django==3.2.4
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pygments==2.9.0
pylint==2.8.3
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
wrapt==1.12.1
